\name{calcDensity}
\alias{calcDensity}
\title{Calculate density of html text to overall length of html tree text...}
\usage{calcDensity(xn, annotate=TRUE)
}
\description{Calculate density of html text to overall length of html tree text}
\author{Mario Annau}
\seealso{\code{\link{extractContentDOM}}, \code{\link{xmlNode}}}
\arguments{\item{xn}{object of class xmlNode}
\item{annotate}{Specifies if \code{xn} should be annotated, defaults to TRUE}
}

