\name{ListSource}
\alias{ListSource}
\title{ListSource}
\usage{ListSource(x, encoding="UTF-8")
}
\description{Convert named list to a corpus}
\author{Mario Annau}
\alias{ListSource}
\alias{readList}
\arguments{\item{x}{named list to be converted}
\item{encoding}{Encoding to be used for source reading}
}

