\name{listtoxml}
\alias{listtoxml}
\title{Convert contents of a named list to an XML index...}
\usage{listtoxml(corpuslist, idxfilename)
}
\description{Convert contents of a named list to an XML index}
\note{This function is still in an experimental status}
\arguments{\item{corpuslist}{list holding elements of a (corpus) list}
\item{idxfilename}{character specifying full path of filename of XML index to be created}
}

